Refactor bookmark handling and improve type safety in cosmic journey components
